DG_AI/requirements.txt
# DrainGuard AI – Python Dependencies
# Install: pip install -r requirements.txt

# Data & Computation
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Serial Communication (for ESP32 hardware)
pyserial>=3.5

# Dashboard
streamlit>=1.28.0
plotly>=5.18.0

# Utilities
python-dateutil>=2.8.0
